requirements.txt
flask
gunicorn
groq
youtube-transcript-api
fpdf2
requests
python-dotenv
